[#57] feat: Remake point component as an atom
