--- src/components/atoms/icons/coinIcon.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Image } from "react-native";
-
-const CoinIcon = ({ style }: any) => {
-  return <Image style={style} source={require("@assets/btn_point.png")} />;
-};
-
-export default CoinIcon;
