+++ src/components/atoms/icons/point.style.ts
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  pointWrapper: {
+    flexDirection: "row"
+  },
+  pointIcon: {
+    resizeMode: "contain",
+    width: 12,
+    height: 12,
+    marginTop: 4,
+    marginRight: 3
+  }
+});
+
+export default styles;
