+++ src/components/atoms/index.ts
@@ -1,6 +1,5 @@
 export { default as Text } from "./text/text";
 export { default as ProfileIcon } from "./icons/profileIcon";
-export { default as CoinIcon } from "./icons/coinIcon";
 export { default as RightArrowIcon } from "./icons/rightArrowIcon";
 export { default as ShopIcon } from "./icons/shopIcon";
 export { default as StorageIcon } from "./icons/storageIcon";
