+++ src/components/molecules/index.ts
@@ -2,7 +2,7 @@ export { default as Aim } from "./main/aim";
 export { default as Banner } from "./main/banner";
 export { default as CircleProgress } from "./main/circleProgress";
 export { default as FloatingButton } from "./main/floatingButton";
-export { default as Point } from "./main/point";
+export { default as Point } from "../atoms/icons/point";
 export { default as RecordsOfMonth } from "./main/recordsOfMonth";
 export { default as Form } from "./auth/form";
 export { default as CheckTermsButton } from "./auth/checkTermsButton";
