--- src/components/molecules/main/point.style.ts
@@ -1,19 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  myPointContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "flex-end"
-  },
-  myPointImg: {
-    width: 12,
-    height: 12,
-    marginRight: 4
-  },
-  myPointFont: {
-    fontSize: 11
-  }
-});
-
-export default styles;
